# Controllers/TodoItemsController.cs
/// <summary>
/// This file is the controller for the TodoItems API. It contains methods for handling HTTP requests to the API endpoints.
/// </summary>

using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using TodoApi.Models;

namespace TodoApi.Controllers
{

    [Route("logo.png")]
    public class LogoController : ControllerBase
    {
        [HttpGet]
        public IActionResult Get()
        {
            var file = System.IO.File.OpenText("logo.png");
            return File(file.BaseStream, "image/png");
        }

    }
       [Route(".well-known/ai-plugin.json")]
    public class AiPluginController : ControllerBase
    {
        [HttpGet]
        public IActionResult Get()
        {
            var file = System.IO.File.OpenText("ai-plugin.json");
            return File(file.BaseStream, "application/json");
        }

    }

    [Route("openapi.yaml")]
    public class OpenApiYamlController : ControllerBase
    {

        [HttpGet]
        public IActionResult Get()
        {
            var file = System.IO.File.OpenText("openapi.yaml");
            return File(file.BaseStream, "application/yaml");
        }
    }

    [Route("api/TodoItems")]
    [ApiController]
    public class TodoItemsController : ControllerBase
    {
        private readonly TodoContext _context;

        public TodoItemsController(TodoContext context)
        {
            _context = context;
        }

        // GET: api/TodoItems
        [HttpGet]
        public async Task<ActionResult<IEnumerable<TodoItem>>> GetTodoItems(string name = null)
        {
            if (name == null)
            {
                return await _context.TodoItems.ToListAsync();
            }
            return _context.TodoItems.Where(x => x.Name.Contains(name)).ToList(); ;
        }

        // GET: api/TodoItems/5
        [HttpGet("{id}")]
        public async Task<ActionResult<TodoItem>> GetTodoItem(int id)
        {
            var todoItem = await _context.TodoItems.FindAsync(id);

            if (todoItem == null)
            {
                return NotFound();
            }

            return todoItem;
        }

        // PUT: api/TodoItems/5
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPut("{id}")]
        public async Task<IActionResult> PutTodoItem(int id, TodoItem todoItem)
        {
            if (id != todoItem.Id)
            {
                return BadRequest();
            }

            _context.Entry(todoItem).State = EntityState.Modified;

            try
            {
                await _context.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!TodoItemExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return NoContent();
        }

        // POST: api/TodoItems
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPost]
        public async Task<ActionResult<TodoItem>> PostTodoItem(TodoItem todoItem)
        {
            _context.TodoItems.Add(todoItem);
            await _context.SaveChangesAsync();

            return CreatedAtAction(nameof(GetTodoItem), new { id = todoItem.Id }, todoItem);
        }

        // DELETE: api/TodoItems/5
        [HttpDelete("{id}")]
        public async Task<IActionResult> DeleteTodoItem(int id)
        {
            var todoItem = await _context.TodoItems.FindAsync(id);
            if (todoItem == null)
            {
                return NotFound();
            }

            _context.TodoItems.Remove(todoItem);
            await _context.SaveChangesAsync();

            return NoContent();
        }

        private bool TodoItemExists(int id)
        {
            return _context.TodoItems.Any(e => e.Id == id);
        }
    }
}